zstd: new package at 1.4.4
Looks like we'll be forced to include zstd in the near future.
Software releases over the past few months have included zstd
support with more software to follow (non-optionally).

The GNU autotools suite has added support for creating dist
tarballs with zstd as the compression method. We may start to
see release tarballs of software compressed with zstd.

ccache will use zstd in the next major release and this is a
mandatory change. It cannot be avoided or disabled. We'd need
zstd to continue to keep ccache up-to-date (important).

GCC 10, GRUB and other software have received optional support
for zstd. These may be made mandatory in the future, it's too
early to tell.

Basically, a mandatory requirement on zstd is coming for some
software. This commit just readies KISS for the future though
I would have preferred to not include it at all (similar
reasoning to Debian).