freetype-harfbuzz: revert to 2.10.2+2.6.8
The swap to meson also causes the libtool .la files to no longer
exist as libtool is only called in the autotools build. We swapped
to meson as upstream will no longer publish release tarballs.

The removal of these files causes any other packages which both
link to harfbuzz and use autotools to fail to build. This package
is being reverted to prevent these issues.

A permanent fix is coming next. All .la files will be removed by
the package manager post build. This will prevent the issue from
ever arising again as well as dropping unneeded files from a
myriad of packages.