chromium: testing at 96.0.4664.45
Currently cannot get to build as patch errors out when trying to apply
the musl patches from Void. I have not tried with the previous musl
patches, but seeing as there is a big difference between the two sets I
doubt they would work. If you are able to get this to compile, make a
PR.